chore: add course description section
