+++ src/components/CourseDetailsComponent/SubComponents/AboutElements/CourseSchedule.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function CourseSchedule() {
+  return <div></div>;
+}
+
+export default CourseSchedule;
